added security to favorites endpoints

+++ src/main/java/com/revature/jash/web/controllers/UserController.java
@@ -83,13 +83,25 @@ public class UserController {
 
     @PostMapping(value="/favorites", produces="application/json")
     @ResponseStatus(HttpStatus.OK)
-    public void addFavorite(@PathVariable String user_id, @PathVariable String collection_id) {
+    public void addFavorite(@PathVariable String user_id, @PathVariable String collection_id, HttpServletRequest req) {
+        Principal principal = parser.parseToken(req).orElseThrow(() -> new AuthenticationException("Request originates from an unauthenticated source."));
+        String requester = principal.getId();
+        String accessed = user_id;
+        if(!requester.equals(accessed)) {
+            throw new UserForbiddenException("Not allowed to update other Users");
+        }
         userService.addFavorite(user_id, collection_id);
     }
 
     @DeleteMapping(value="/favorites", produces="application/json")
     @ResponseStatus(HttpStatus.OK)
-    public void removeFavorite(@PathVariable String user_id, @PathVariable String collection_id) {
+    public void removeFavorite(@PathVariable String user_id, @PathVariable String collection_id, HttpServletRequest req) {
+        Principal principal = parser.parseToken(req).orElseThrow(() -> new AuthenticationException("Request originates from an unauthenticated source."));
+        String requester = principal.getId();
+        String accessed = user_id;
+        if(!requester.equals(accessed)) {
+            throw new UserForbiddenException("Not allowed to update other Users");
+        }
         userService.removeFavorite(user_id, collection_id);
     }
 }
